+++ apple-flat-package/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 Released on ReleaseDate.
 
+* `VolumeCheck`'s `script` field has been changed from `bool` -> `Option<bool>`.
+  (#213)
 * MSRV 1.81 -> 1.88.
 * `scroll` 0.12 -> 0.13.
 * `serde-xml-rs` 0.6. -> 0.8.